+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# unreleased
+
+_Breaking Changes_
+
+- Ignore `DOMException` errors when calling `insertRule` while inserting stylesheet
+
 # 3.0.0
 
 - Add typescript types
@@ -5,6 +11,7 @@
 _Breaking Changes_
 
 - drop IE8 support
+- Throw any errors that are not `SyntaxError` when calling `insertRule` while inserting stylesheet
 
 # 2.0.0